--- packages/visual-editor/src/components/Phone.test.tsx
@@ -1,103 +1,0 @@
-import * as React from "react";
-import { describe, it, expect } from "vitest";
-import { axe, viewports } from "./WCAG/WCAG.setup.ts";
-import { render as reactRender } from "@testing-library/react";
-import { Phone, VisualEditorProvider } from "@yext/visual-editor";
-import { Render, Config } from "@measured/puck";
-import { page } from "@vitest/browser/context";
-
-describe.each(viewports)("Phone $name", ({ width, height }) => {
-  const puckConfig: Config = {
-    components: { Phone },
-    root: {
-      render: ({ children }) => {
-        return <>{children}</>;
-      },
-    },
-  };
-
-  it("should pass wcag with default props", async () => {
-    const { container } = reactRender(
-      <VisualEditorProvider templateProps={{ document: {} }}>
-        <Render
-          config={puckConfig}
-          data={{
-            content: [
-              {
-                type: "Phone",
-                props: { id: "abc", ...Phone.defaultProps },
-              },
-            ],
-          }}
-        />
-      </VisualEditorProvider>
-    );
-
-    await page.viewport(width, height);
-    await page.screenshot();
-    const results = await axe(container);
-    expect(results).toHaveNoViolations();
-  });
-
-  it("should pass wcag with data - no hyperlink", async () => {
-    const { container } = reactRender(
-      <VisualEditorProvider templateProps={{ document: {} }}>
-        <Render
-          config={puckConfig}
-          data={{
-            content: [
-              {
-                type: "Phone",
-                props: {
-                  id: "abc",
-                  phone: {
-                    field: "mainPhone",
-                    constantValue: "+12025551010",
-                    constantValueEnabled: true,
-                  },
-                  includeHyperlink: false,
-                },
-              },
-            ],
-          }}
-        />
-      </VisualEditorProvider>
-    );
-
-    await page.viewport(width, height);
-    await page.screenshot();
-    const results = await axe(container);
-    expect(results).toHaveNoViolations();
-  });
-
-  it("should pass wcag with data - hyperlink", async () => {
-    const { container } = reactRender(
-      <VisualEditorProvider templateProps={{ document: {} }}>
-        <Render
-          config={puckConfig}
-          data={{
-            content: [
-              {
-                type: "Phone",
-                props: {
-                  id: "abc",
-                  phone: {
-                    field: "mainPhone",
-                    constantValue: "+12025551010",
-                    constantValueEnabled: true,
-                  },
-                  includeHyperlink: true,
-                },
-              },
-            ],
-          }}
-        />
-      </VisualEditorProvider>
-    );
-
-    await page.viewport(width, height);
-    await page.screenshot();
-    const results = await axe(container);
-    expect(results).toHaveNoViolations();
-  });
-});
