+++ packages/visual-editor/src/components/contentBlocks/HoursTable.tsx
