+++ packages/visual-editor/src/components/contentBlocks/Image.tsx
