+++ packages/visual-editor/src/components/contentBlocks/MapboxStaticMap.tsx
