+++ packages/visual-editor/src/components/contentBlocks/TextList.test.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
 import { describe, it, expect } from "vitest";
-import { axe, testHours, viewports } from "./WCAG/WCAG.setup.ts";
+import { axe, viewports } from "../WCAG/WCAG.setup.ts";
 import { render as reactRender } from "@testing-library/react";
-import { HoursTable, VisualEditorProvider } from "@yext/visual-editor";
+import { TextList, VisualEditorProvider } from "@yext/visual-editor";
 import { Render, Config } from "@measured/puck";
 import { page } from "@vitest/browser/context";
 
-describe.each(viewports)("HoursTable $name", ({ width, height }) => {
+describe.each(viewports)("TextList $name", ({ width, height }) => {
   const puckConfig: Config = {
-    components: { HoursTable },
+    components: { TextList },
     root: {
       render: ({ children }) => {
         return <>{children}</>;
@@ -18,21 +18,14 @@ describe.each(viewports)("HoursTable $name", ({ width, height }) => {
 
   it("should pass wcag with default props", async () => {
     const { container } = reactRender(
-      <VisualEditorProvider
-        templateProps={{
-          document: {
-            hours: testHours,
-            additionalHoursText: "Hours Text",
-          },
-        }}
-      >
+      <VisualEditorProvider templateProps={{ document: {} }}>
         <Render
           config={puckConfig}
           data={{
             content: [
               {
-                type: "HoursTable",
-                props: { id: "abc", ...HoursTable.defaultProps },
+                type: "TextList",
+                props: { id: "abc", ...TextList.defaultProps },
               },
             ],
           }}
@@ -46,20 +39,23 @@ describe.each(viewports)("HoursTable $name", ({ width, height }) => {
     expect(results).toHaveNoViolations();
   });
 
-  it("should pass wcag with empty document", async () => {
+  it("should pass wcag with data", async () => {
     const { container } = reactRender(
-      <VisualEditorProvider
-        templateProps={{
-          document: {},
-        }}
-      >
+      <VisualEditorProvider templateProps={{ document: {} }}>
         <Render
           config={puckConfig}
           data={{
             content: [
               {
-                type: "HoursTable",
-                props: { id: "abc", ...HoursTable.defaultProps },
+                type: "TextList",
+                props: {
+                  id: "abc",
+                  list: {
+                    field: "",
+                    constantValue: ["abc", "def"],
+                    constantValueEnabled: true,
+                  },
+                },
               },
             ],
           }}
