+++ packages/visual-editor/src/components/contentBlocks/index.ts
@@ -1,0 +1,12 @@
+export { Address, type AddressProps } from "./Address.tsx";
+export { BodyText, type BodyTextProps } from "./BodyText.tsx";
+export { CTAWrapper, type CTAWrapperProps } from "./CtaWrapper.tsx";
+export { Emails, type EmailsProps } from "./Emails.tsx";
+export { GetDirections, type GetDirectionsProps } from "./GetDirections.tsx";
+export { HeadingText, type HeadingTextProps } from "./HeadingText.tsx";
+export { HoursStatus, type HoursStatusProps } from "./HoursStatus.tsx";
+export { HoursTable, type HoursTableProps } from "./HoursTable.tsx";
+export { ImageWrapper, type ImageWrapperProps } from "./Image.tsx";
+export { MapboxStaticMap, type MapboxStaticProps } from "./MapboxStaticMap.tsx";
+export { Phone, type PhoneProps } from "./Phone.tsx";
+export { TextList, type TextListProps } from "./TextList.tsx";
