+++ packages/visual-editor/src/components/index.ts
@@ -1,46 +1,10 @@
 export * from "./atoms/index.ts";
 export * from "./_componentCategories.ts";
 export * from "./cards/index.ts";
-export { Address, type AddressProps } from "./Address.tsx";
-export { BannerSection, type BannerSectionProps } from "./Banner.tsx";
-export { BodyText, type BodyTextProps } from "./BodyText.tsx";
-export {
-  BreadcrumbsSection,
-  type BreadcrumbsSectionProps,
-} from "./Breadcrumbs.tsx";
-export {
-  CoreInfoSection,
-  type CoreInfoSectionProps,
-} from "./CoreInfoSection.tsx";
-export { CTAWrapper, type CTAWrapperProps } from "./CtaWrapper.tsx";
+export * from "./contentBlocks/index.ts";
+export * from "./layoutBlocks/index.ts";
+export * from "./pageSections/index.ts";
 export { Directory, type DirectoryProps } from "./Directory.tsx";
-export { Emails, type EmailsProps } from "./Emails.tsx";
-export { FAQsSection, type FAQsSectionProps } from "./FAQs.tsx";
-export { Flex, type FlexProps } from "./Flex.tsx";
 export { Footer, type FooterProps } from "./Footer.tsx";
-export { GetDirections, type GetDirectionsProps } from "./GetDirections.tsx";
-export { Grid, type GridProps } from "./Grid.tsx";
 export { Header, type HeaderProps } from "./Header.tsx";
-export { HeadingText, type HeadingTextProps } from "./HeadingText.tsx";
-export { HeroSection, type HeroSectionProps } from "./HeroSection.tsx";
-export { HoursStatus, type HoursStatusProps } from "./HoursStatus.tsx";
-export { HoursTable, type HoursTableProps } from "./HoursTable.tsx";
-export { ImageWrapper, type ImageWrapperProps } from "./Image.tsx";
-export { MapboxStaticMap, type MapboxStaticProps } from "./MapboxStaticMap.tsx";
-export {
-  NearbyLocationsSection,
-  type NearbyLocationsSectionProps,
-} from "./NearbyLocations.tsx";
-export { People, type PeopleProps } from "./People.tsx";
-export { Phone, type PhoneProps } from "./Phone.tsx";
-export {
-  PhotoGallerySection,
-  type PhotoGallerySectionProps,
-} from "./PhotoGallerySection.tsx";
-export { PromoSection, type PromoSectionProps } from "./Promo.tsx";
-export {
-  SectionContainer,
-  type SectionContainerProps,
-} from "./SectionContainer.tsx";
-export { TextList, type TextListProps } from "./TextList.tsx";
 export { defaultThemeConfig } from "./DefaultThemeConfig.ts";
