+++ packages/visual-editor/src/components/layoutBlocks/index.ts
@@ -1,0 +1,3 @@
+export { Collection, type CollectionProps } from "./Collection.tsx";
+export { Flex, type FlexProps } from "./Flex.tsx";
+export { Grid, type GridProps } from "./Grid.tsx";
