+++ packages/visual-editor/src/components/pageSections/CoreInfoSection.tsx
