+++ packages/visual-editor/src/components/pageSections/FAQs.test.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
 import { describe, it, expect } from "vitest";
-import { axe, viewports } from "./WCAG/WCAG.setup.ts";
+import { axe, viewports } from "../WCAG/WCAG.setup.ts";
 import { render as reactRender } from "@testing-library/react";
-import { HeadingText, VisualEditorProvider } from "@yext/visual-editor";
+import { FAQsSection, VisualEditorProvider } from "@yext/visual-editor";
 import { Render, Config } from "@measured/puck";
 import { page } from "@vitest/browser/context";
 
-describe.each(viewports)("HeadingText $name", ({ width, height }) => {
+describe.each(viewports)("FAQsSection $name", ({ width, height }) => {
   const puckConfig: Config = {
-    components: { HeadingText },
+    components: { FAQsSection },
     root: {
       render: ({ children }) => {
         return <>{children}</>;
@@ -24,8 +24,8 @@ describe.each(viewports)("HeadingText $name", ({ width, height }) => {
           data={{
             content: [
               {
-                type: "HeadingText",
-                props: { id: "abc", ...HeadingText.defaultProps },
+                type: "FAQsSection",
+                props: { id: "abc", ...FAQsSection.defaultProps },
               },
             ],
           }}
