+++ packages/visual-editor/src/components/pageSections/HeroSection.tsx
@@ -18,7 +18,10 @@ import {
   PageSection,
   YextField,
 } from "@yext/visual-editor";
-import { ImageWrapperFields, resolvedImageFields } from "./Image.js";
+import {
+  ImageWrapperFields,
+  resolvedImageFields,
+} from "../contentBlocks/Image.js";
 
 const PLACEHOLDER_IMAGE_URL = "https://placehold.co/640x360";
