+++ packages/visual-editor/src/components/pageSections/index.ts
@@ -1,0 +1,24 @@
+export { BannerSection, type BannerSectionProps } from "./Banner.tsx";
+export {
+  BreadcrumbsSection,
+  type BreadcrumbsSectionProps,
+} from "./Breadcrumbs.tsx";
+export {
+  CoreInfoSection,
+  type CoreInfoSectionProps,
+} from "./CoreInfoSection.tsx";
+export { FAQsSection, type FAQsSectionProps } from "./FAQs.tsx";
+export { HeroSection, type HeroSectionProps } from "./HeroSection.tsx";
+export {
+  NearbyLocationsSection,
+  type NearbyLocationsSectionProps,
+} from "./NearbyLocations.tsx";
+export {
+  PhotoGallerySection,
+  type PhotoGallerySectionProps,
+} from "./PhotoGallerySection.tsx";
+export { PromoSection, type PromoSectionProps } from "./Promo.tsx";
+export {
+  SectionContainer,
+  type SectionContainerProps,
+} from "./SectionContainer.tsx";
